todoapi/views.py
from rest_framework import generics

from .models import Task
from .permissions import IsOwnerOrReadOnly
from .serializer import TaskSerializer


class TasksList(generics.ListCreateAPIView):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer


class TasksDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer
    permission_classes = (IsOwnerOrReadOnly,)
