pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["Odeh <abuzaid163@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
psycopg2-binary = "^2.9.2"
asgiref = "^3.4.1"
pytz = "^2021.3"
sqlparse = "^0.4.2"
djangorestframework = "^3.12.4"


[tool.poetry.dev-dependencies]
black = "^21.11b1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
